+++ component/doc.go
@@ -1,2 +1,13 @@
+//
+//
+// How to create custom Components
+//
+// By convention if a Component (lets call it Box) needs Properties the package
+// should have a struct called BoxProps.
+//
+//
+// As a general starting point many of the core Components are designed to
+// be reference implementations of retort's features.
+//
 // package component has common low level components for you to build with
 package component // import "retort.dev/component"
