+++ r/useQuit.go
@@ -1,0 +1,9 @@
+package r
+
+// UseQuit returns a single function that when invoked
+// will exit the application
+func UseQuit() func() {
+	return func() {
+		close(quitChan)
+	}
+}
